NRFFOSDK-10937: Change the type of transfer lengths to size_t
Type of transfer lengths was unified for SPI, TWI, TWIM, UART,
and UARTE drivers. Added checking if the lengths are not exceeding
the limits for peripherals using EasyDMA.

The function "nrfx_spi_transfer" has been removed as "nrfx_spi_xfer"
fully covers its functionality.

+++ drivers/include/nrfx_twi.h
@@ -100,8 +100,8 @@ typedef struct
 {
     nrfx_twi_xfer_type_t    type;             ///< Type of transfer.
     uint8_t                 address;          ///< Slave address.
-    uint32_t                primary_length;   ///< Number of bytes transferred.
-    uint32_t                secondary_length; ///< Number of bytes transferred.
+    size_t                  primary_length;   ///< Number of bytes transferred.
+    size_t                  secondary_length; ///< Number of bytes transferred.
     uint8_t *               p_primary_buf;    ///< Pointer to transferred data.
     uint8_t *               p_secondary_buf;  ///< Pointer to transferred data.
 } nrfx_twi_xfer_desc_t;
@@ -226,7 +226,7 @@ void nrfx_twi_disable(nrfx_twi_t const * p_instance);
 nrfx_err_t nrfx_twi_tx(nrfx_twi_t const * p_instance,
                        uint8_t            address,
                        uint8_t const *    p_data,
-                       uint32_t           length,
+                       size_t             length,
                        bool               no_stop);
 
 /**
@@ -250,7 +250,7 @@ nrfx_err_t nrfx_twi_tx(nrfx_twi_t const * p_instance,
 nrfx_err_t nrfx_twi_rx(nrfx_twi_t const * p_instance,
                        uint8_t            address,
                        uint8_t *          p_data,
-                       uint32_t           length);
+                       size_t             length);
 
 /**
  * @brief Function for preparing a TWI transfer.
@@ -304,7 +304,7 @@ bool nrfx_twi_is_busy(nrfx_twi_t const * p_instance);
  *
  * @return     Data count.
  */
-uint32_t nrfx_twi_data_count_get(nrfx_twi_t const * const p_instance);
+size_t nrfx_twi_data_count_get(nrfx_twi_t const * const p_instance);
 
 /**
  * @brief Function for returning the address of a STOPPED TWI event.
